wandb/run-20260413_150307-jto4rtk0/files/requirements.txt
gto==1.9.0
sympy==1.14.0
sentry-sdk==2.57.0
ipython==9.12.0
jupyterlab==4.5.6
jsonschema==4.26.0
pydantic_core==2.41.5
pyOpenSSL==26.0.0
nvidia-cusparse-cu12==12.5.8.93
multidict==6.7.1
gssapi==1.11.1
ptyprocess==0.7.0
pyparsing==3.3.2
nvidia-cudnn-cu12==9.10.2.21
jupyter-console==6.6.3
jupyterlab_pygments==0.3.0
click-plugins==1.1.1.2
Jinja2==3.1.6
pycparser==2.22
jupyter_server_terminals==0.5.4
argon2-cffi-bindings==25.1.0
rfc3987-syntax==1.1.0
jupyter==1.1.1
anyio==4.13.0
zc.lockfile==4.0
nvidia-cublas-cu12==12.8.4.1
fqdn==1.5.1
dvc-render==1.0.2
webcolors==25.10.0
nvidia-cuda-runtime-cu12==12.8.90
funcy==2.0
babel==2.18.0
billiard==4.2.4
mdurl==0.1.2
click-didyoumean==0.3.1
arrow==1.4.0
nest_asyncio==1.6.0
mpmath==1.3.0
backports.zstd==1.3.0
dvc-studio-client==0.22.0
certifi==2026.2.25
vine==5.1.0
dvc-data==3.18.3
PyYAML==6.0.3
comm==0.2.3
MarkupSafe==3.0.3
jmespath==1.1.0
requests==2.33.1
click-repl==0.3.0
typing-inspection==0.4.2
GitPython==3.1.46
entrypoints==0.4
voluptuous==0.16.0
appdirs==1.4.4
eval_type_backport==0.3.1
rfc3339_validator==0.1.4
typing_utils==0.1.0
executing==2.2.1
importlib_resources==7.0.0
Brotli==1.2.0
pydot==4.0.1
aiosignal==1.4.0
wandb==0.25.1
nvidia-nvtx-cu12==12.8.90
semver==3.0.4
tomli==2.4.1
isoduration==20.11.0
rpds-py==0.30.0
docker-pycreds==0.4.0
dvc-task==0.40.2
aiohttp-retry==2.9.1
jupyterlab_widgets==3.0.16
pydantic==2.12.5
tzdata==2026.1
sqltrie==0.11.2
stack_data==0.6.3
exceptiongroup==1.3.1
cffi==2.0.0
nvidia-cuda-cupti-cu12==12.8.90
typing_extensions==4.15.0
parso==0.8.6
protobuf==6.33.5
charset-normalizer==3.4.7
CompactReasoningModels==0.1.0
json5==0.14.0
importlib_metadata==8.8.0
propcache==0.3.1
gitdb==4.0.12
notebook==7.5.5
shortuuid==1.0.13
pygit2==1.19.2
pandas==3.0.2
markdown-it-py==4.0.0
diskcache==5.6.3
hyperframe==6.1.0
tinycss2==1.4.0
pygtrie==2.5.0
kombu==5.5.4
jupyter_client==8.8.0
boto3==1.42.88
httpx==0.28.1
dvc-http==2.32.0
nvidia-cufile-cu12==1.13.1.3
asyncssh==2.22.0
aiohttp==3.13.5
smmap==5.0.3
ipykernel==7.2.0
annotated-doc==0.0.4
jupyter_core==5.9.1
async-lru==2.3.0
pydantic-settings==2.13.1
overrides==7.7.0
typer==0.24.1
jupyterlab_server==2.28.0
beautifulsoup4==4.14.3
ipython_pygments_lexers==1.1.1
httpcore==1.0.9
frozenlist==1.7.0
cached-property==1.5.2
nvidia-curand-cu12==10.3.9.90
argon2-cffi==25.1.0
ruamel.yaml==0.19.1
annotated-types==0.7.0
dvc==3.67.1
sniffio==1.3.1
python-dateutil==2.9.0.post0
nvidia-nccl-cu12==2.27.5
python-dotenv==1.2.2
numpy==2.4.3
Pygments==2.20.0
amqp==5.3.1
h11==0.16.0
pure_eval==0.2.3
referencing==0.37.0
asttokens==3.0.1
click==8.3.2
scmrepo==3.6.2
wcwidth==0.6.0
pexpect==4.9.0
attrs==26.1.0
nvidia-cusolver-cu12==11.7.3.90
widgetsnbextension==4.0.15
soupsieve==2.8.3
platformdirs==4.9.6
jedi==0.19.2
orjson==3.11.8
mistune==3.2.0
PySocks==1.7.1
dpath==2.2.0
antlr4-python3-runtime==4.9.3
zipp==3.23.0
hydra-core==1.3.2
traitlets==5.14.3
iterative-telemetry==0.0.10
setproctitle==1.3.7
nbformat==5.10.4
grandalf==0.7
idna==3.11
cuda-bindings==12.9.4
atpublic==7.0.0
colorama==0.4.6
setuptools==82.0.1
tomlkit==0.14.0
yarl==1.23.0
rich==15.0.0
tqdm==4.67.3
nvidia-cusparselt-cu12==0.7.1
distro==1.9.0
fastjsonschema==2.21.2
triton==3.6.0
matplotlib-inline==0.2.1
cryptography==46.0.7
nvidia-nvshmem-cu12==3.4.5
flufl.lock==9.0.0
jupyter_server==2.17.0
networkx==3.6.1
ipywidgets==8.1.8
bleach==6.3.0
defusedxml==0.7.1
six==1.17.0
rfc3986-validator==0.1.1
decorator==5.2.1
botocore==1.42.88
dictdiffer==0.9.0
celery==5.5.3
nvidia-nvjitlink-cu12==12.8.93
configobj==5.0.9
pandocfilters==1.5.0
webencodings==0.5.1
shtab==1.8.0
dulwich==1.0.0
ruamel.yaml.clib==0.2.15
packaging==26.0
notebook_shim==0.2.4
jupyter-events==0.12.0
pathspec==1.0.4
pyzmq==27.1.0
s3transfer==0.16.0
nvidia-cuda-nvrtc-cu12==12.8.93
urllib3==2.6.3
filelock==3.25.2
flatten-dict==0.4.2
nbconvert==7.17.1
pathlib2==2.3.7.post1
terminado==0.18.1
h2==4.3.0
jsonpointer==3.1.1
dvc-objects==5.2.0
jsonschema-specifications==2025.9.1
nvidia-cufft-cu12==11.3.3.83
hpack==4.1.0
shellingham==1.5.4
nbclient==0.10.4
torch==2.10.0+cu128
cuda-pathfinder==1.5.2
debugpy==1.8.20
lark==1.3.1
websocket-client==1.9.0
omegaconf==2.3.0
tornado==6.5.5
fsspec==2026.3.0
prometheus_client==0.25.0
future==1.0.0
prompt_toolkit==3.0.52
python-json-logger==2.0.7
tabulate==0.10.0
jupyter-lsp==2.3.1
uri-template==1.3.0
aiohappyeyeballs==2.6.1
psutil==7.2.2
Send2Trash==2.1.0
